requirements.txt
elasticsearch==2.3.0
jsonklog==0.15.0
six==1.10.0
python-logstash==0.4.6
fluent-logger==0.9.3
PyJWT==1.5.3
requests==2.19.1 # 2.20.* имеют баг с докачкой большого потока
redis==2.10.6
kafka-python==2.0.1